Fix printing wrong line when failing to parse an item
When trying to parse items in bulk, we advance the iterator to the next
line and save it as the current line variable. This variable is printed
in cases of failures when parsing. Thus, the next line will be printed
instead of the actual current line. We can solve this by just storing
the next line in another variable.